pyproject.toml
[tool.poetry]
name = "your-pdf-qna"
version = "0.1.0"
description = ""
authors = ["Rajesh Rajagopalan <rrajesh1979@gmail.com>"]
readme = "README.md"
packages = [{include = "your_pdf_qna"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
streamlit = "^1.12.0"
langchain = "^0.0.159"
tiktoken = "^0.3.3"
ipython = "^8.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

